DOCUMENTATION: Better npm run doc, plus minor folder reorganization (#693)
