+++ src/Documentation/doc/advanced/intelligence.md
