+++ src/Documentation/doc/advanced/stanek.md
