+++ src/Documentation/doc/basic/terminal.md
