+++ src/Documentation/doc/programming/hackingalgorithms.md
